requirements.txt
asgiref==3.9.1
cryptography==45.0.5
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
psutil==7.0.0
PyMySQL==1.1.1
python-dotenv==1.1.1
requests==2.32.4
SQLAlchemy==2.0.41
uvicorn==0.35.0
